+++ src/main/java/model/BankAccount.java
@@ -3,6 +3,6 @@ package model;
 import java.util.Queue;
 
 public class BankAccount {
-    public int balance;
+    public double balance;
     public Queue<AccountStatement> accountStatements;
 }
