+++ src/main/java/service/BankStatementService.java
@@ -7,7 +7,7 @@ import java.time.ZonedDateTime;
 public interface BankStatementService {
 
 
-     void updateAccountStatement(BankAccount account, int amount, String operation, ZonedDateTime zonedDateTime);
+     void updateAccountStatement(BankAccount account, double amount, String operation, ZonedDateTime zonedDateTime);
 
      void printAccountStatement(BankAccount bankAccount);
 }
