+++ src/main/java/service/StatementService.java
@@ -9,7 +9,7 @@ public class StatementService implements BankStatementService {
 
 
     @Override
-    public void updateAccountStatement(BankAccount account, int amount, String operation, ZonedDateTime zonedDateTime) {
+    public void updateAccountStatement(BankAccount account, double amount, String operation, ZonedDateTime zonedDateTime) {
         account.accountStatements.add(new AccountStatement(zonedDateTime, amount, operation));
     }
