+++ src/main/java/service/WithdrawalOperations.java
@@ -4,6 +4,6 @@ import model.BankAccount;
 
 public interface WithdrawalOperations {
 
-    void withdrawal(BankAccount account, int amount);
+    void withdrawal(BankAccount account, double amount);
 
 }
